refactor github controllers
