+++ backend/controllers/github_setup.go
@@ -1,0 +1,163 @@
+package controllers
+
+import (
+	"context"
+	"encoding/base64"
+	"encoding/json"
+	"fmt"
+	"github.com/diggerhq/digger/backend/utils"
+	"github.com/gin-gonic/gin"
+	"log/slog"
+	"math/rand"
+	"net/http"
+	"net/url"
+	"os"
+)
+
+func GithubAppSetup(c *gin.Context) {
+	slog.Info("Setting up GitHub app")
+
+	type githubWebhook struct {
+		URL    string `json:"url"`
+		Active bool   `json:"active"`
+	}
+
+	type githubAppRequest struct {
+		Description           string            `json:"description"`
+		Events                []string          `json:"default_events"`
+		Name                  string            `json:"name"`
+		Permissions           map[string]string `json:"default_permissions"`
+		Public                bool              `json:"public"`
+		RedirectURL           string            `json:"redirect_url"`
+		CallbackUrls          []string          `json:"callback_urls"`
+		RequestOauthOnInstall bool              `json:"request_oauth_on_install"`
+		SetupOnUpdate         bool              `json:"setup_on_update"`
+		URL                   string            `json:"url"`
+		Webhook               *githubWebhook    `json:"hook_attributes"`
+	}
+
+	host := os.Getenv("HOSTNAME")
+	manifest := &githubAppRequest{
+		Name:        fmt.Sprintf("Digger app %v", rand.Int31()),
+		Description: fmt.Sprintf("Digger hosted at %s", host),
+		URL:         host,
+		RedirectURL: fmt.Sprintf("%s/github/exchange-code", host),
+		Public:      false,
+		Webhook: &githubWebhook{
+			Active: true,
+			URL:    fmt.Sprintf("%s/github-app-webhook", host),
+		},
+		CallbackUrls:          []string{fmt.Sprintf("%s/github/callback", host)},
+		SetupOnUpdate:         true,
+		RequestOauthOnInstall: true,
+		Events: []string{
+			"check_run",
+			"create",
+			"delete",
+			"issue_comment",
+			"issues",
+			"status",
+			"pull_request_review_thread",
+			"pull_request_review_comment",
+			"pull_request_review",
+			"pull_request",
+			"push",
+		},
+		Permissions: map[string]string{
+			"actions":          "write",
+			"contents":         "write",
+			"issues":           "write",
+			"pull_requests":    "write",
+			"repository_hooks": "write",
+			"statuses":         "write",
+			"administration":   "read",
+			"checks":           "write",
+			"members":          "read",
+			"workflows":        "write",
+		},
+	}
+
+	githubHostname := utils.GetGithubHostname()
+	url := &url.URL{
+		Scheme: "https",
+		Host:   githubHostname,
+		Path:   "/settings/apps/new",
+	}
+
+	// https://developer.github.com/apps/building-github-apps/creating-github-apps-using-url-parameters/#about-github-app-url-parameters
+	githubOrg := os.Getenv("GITHUB_ORG")
+	if githubOrg != "" {
+		url.Path = fmt.Sprintf("organizations/%s%s", githubOrg, url.Path)
+		slog.Debug("Using organization-specific GitHub app setup URL", "organization", githubOrg)
+	}
+
+	jsonManifest, err := json.MarshalIndent(manifest, "", " ")
+	if err != nil {
+		slog.Error("Failed to serialize manifest", "error", err)
+		c.Error(fmt.Errorf("failed to serialize manifest %s", err))
+		return
+	}
+
+	slog.Info("Rendering GitHub app setup template",
+		"targetUrl", url.String(),
+		"appName", manifest.Name,
+	)
+
+	c.HTML(http.StatusOK, "github_setup.tmpl", gin.H{"Target": url.String(), "Manifest": string(jsonManifest)})
+}
+
+// GithubSetupExchangeCode handles the user coming back from creating their app
+// A code query parameter is exchanged for this app's ID, key, and webhook_secret
+// Implements https://developer.github.com/apps/building-github-apps/creating-github-apps-from-a-manifest/#implementing-the-github-app-manifest-flow
+func (d DiggerController) GithubSetupExchangeCode(c *gin.Context) {
+	code := c.Query("code")
+	if code == "" {
+		slog.Warn("Missing code query parameter in GitHub setup callback")
+		c.Error(fmt.Errorf("Ignoring callback, missing code query parameter"))
+		return
+	}
+
+	slog.Info("Exchanging code for GitHub app credentials", "code", code)
+
+	// TODO: to make tls verification configurable for debug purposes
+	//var transport *http.Transport = nil
+	//_, exists := os.LookupEnv("DIGGER_GITHUB_SKIP_TLS")
+	//if exists {
+	//	transport = &http.Transport{
+	//		TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
+	//	}
+	//}
+
+	client, err := d.GithubClientProvider.NewClient(nil)
+	if err != nil {
+		slog.Error("Could not create GitHub client", "error", err)
+		c.Error(fmt.Errorf("could not create github client: %v", err))
+		return
+	}
+
+	cfg, _, err := client.Apps.CompleteAppManifest(context.Background(), code)
+	if err != nil {
+		slog.Error("Failed to exchange code for GitHub app", "error", err)
+		c.Error(fmt.Errorf("Failed to exchange code for github app: %s", err))
+		return
+	}
+
+	slog.Info("Successfully retrieved GitHub app credentials",
+		"appName", *cfg.Name,
+		"appId", cfg.GetID(),
+	)
+
+	PEM := cfg.GetPEM()
+	PemBase64 := base64.StdEncoding.EncodeToString([]byte(PEM))
+	c.HTML(http.StatusOK, "github_setup.tmpl", gin.H{
+		"Target":        "",
+		"Manifest":      "",
+		"ID":            cfg.GetID(),
+		"ClientID":      cfg.GetClientID(),
+		"ClientSecret":  cfg.GetClientSecret(),
+		"Key":           PEM,
+		"KeyBase64":     PemBase64,
+		"WebhookSecret": cfg.GetWebhookSecret(),
+		"URL":           cfg.GetHTMLURL(),
+	})
+}
